CHROMIUM: Fix ma-usb driver timeout
Fixes log spam:
schedule_timeout: wrong timeout value ffffffffffffffff
Workqueue: recv_workq __mausb_ip_recv_work [mausb_host]

BUG=b:287173455
TEST=Hotplug and cold boot with ma-usb display. No
more error messages in the kernel log.
UPSTREAM-TASK=b:262748749

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4660481
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>
Tested-by: Stefan Adolfsson <sadolfsson@chromium.org>
Auto-Submit: Stefan Adolfsson <sadolfsson@chromium.org>

+++ drivers/usb/host/mausb/ip_link.c
@@ -23,7 +23,7 @@ static void __mausb_ip_set_options(struct socket *sock, bool udp)
 	lock_sock_nested(sk, 0);
 
 	// Half a second timeout.
-	sk->sk_rcvtimeo = min_t(int, DIV_ROUND_UP(HZ, 2), MAX_SCHEDULE_TIMEOUT);
+	sk->sk_rcvtimeo = min_t(long, DIV_ROUND_UP(HZ, 2), MAX_SCHEDULE_TIMEOUT);
 
 	val = min_t(u32, 0, sysctl_wmem_max);
 	val = min_t(int, val, INT_MAX / 2);
